server/requirements.txt
fastapi
uvicorn
livekit
python-dotenv
